Update https.js
Fixes issues with instances when a server responds with a header indicating the stream is compressed but it is not (usually due to proxy APIs or something similar). This change catches the error thrown by the zlib library and continues with processing the stream. Reported in https://github.com/axios/axios/issues/2406